# swagger/demo-vascomm-store.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "demo-vascomm-store",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/products": {
      "get": {
        "summary": "List",
        "x-apidog-folder": "Admin/Product",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "Admin/Product"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "example": "1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "",
            "required": false,
            "example": "5",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "",
            "required": false,
            "example": "Ed",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "description": "",
            "required": false,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/468090/apis/api-5729177-run",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/product/{id}": {
      "get": {
        "summary": "Show",
        "x-apidog-folder": "Admin/Product",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "Admin/Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/468090/apis/api-5729178-run",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "summary": "Update",
        "x-apidog-folder": "Admin/Product",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "Admin/Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "19",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_method",
            "in": "query",
            "description": "",
            "required": false,
            "example": "PUT",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": false,
            "example": "multipart/form-data",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "nama"
                  },
                  "price": {
                    "type": "string",
                    "example": "50000"
                  },
                  "image": {
                    "type": "string",
                    "example": [],
                    "format": "binary"
                  },
                  "is_active": {
                    "type": "string",
                    "example": "0"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/468090/apis/api-5729234-run",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "summary": "Delete",
        "x-apidog-folder": "Admin/Product",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "Admin/Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "15",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/468090/apis/api-5729423-run",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/client/products": {
      "get": {
        "summary": "List",
        "x-apidog-folder": "Client",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "Client"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "example": "1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "",
            "required": false,
            "example": "5",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "",
            "required": false,
            "example": "id,desc",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "",
            "required": false,
            "example": "nihil",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/468090/apis/api-5734022-run",
        "security": []
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://api.demo-vascomm-store.lndo.site/api",
      "description": "demo-vascomm-store-api"
    }
  ]
}